pyproject.toml
[project]
name = "jane-street-dormant-puzzle"
version = "0.1.0"
description = "Tools for solving Jane Street's Dormant LLM Puzzle — backdoor trigger discovery via linear probes and activation analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "transformers>=4.37.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",

    # MLX backend (Apple Silicon fast inference)
    "mlx-lm>=0.30.0",

    # Puzzle API client
    "jsinfer>=0.2.0",

    # Scientific computing
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",

    # Data handling
    "pandas>=2.1.0",
    "h5py>=3.10.0",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",

    # Experiment tracking & utils
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",

    # Jupyter
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "ipdb>=0.13.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
